src/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyascii==0.3.2
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
bitarray==2.7.3
bleach==6.0.0
blis==0.7.9
boto3==1.24.24
botocore==1.27.24
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
clean-text==0.6.0
cleo==2.0.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
colorama==0.4.6
comm==0.1.3
comtypes==1.1.14
confection==0.0.4
contourpy==1.0.7
coqpit==0.0.17
crashtest==0.4.1
cryptography==40.0.2
cssselect==1.2.0
cupy-cuda12x==12.0.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.28
dateparser==1.1.7
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.15.1
dill==0.3.6
distlib==0.3.6
docker==6.0.1
docopt==0.6.2
duckduckgo-search==2.8.5
dulwich==0.21.5
emoji==1.7.0
encodec==0.1.1
entrypoints==0.4
espeakng==1.0.2
executing==1.2.0
fairseq==0.12.2
fastjsonschema==2.16.3
fastrlock==0.8.1
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.10.7
Flask==2.2.5
Flask-Cors==3.0.10
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
funcy==2.0
future==0.18.3
g2pkk==0.1.2
gevent==22.10.2
gin==0.1.6
gin-config==0.5.0
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gTTS==2.2.4
h5py==3.8.0
html5lib==1.1
httplib2==0.22.0
httpx==0.23.3
hydra-core==1.0.7
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
inflect==5.6.0
installer==0.7.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jamo==0.4.1
jaraco.classes==3.2.3
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.14
jsonlines==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.1.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
keyboard==0.13.5
keyring==23.13.1
kiwisolver==1.4.4
langcodes==3.3.0
librosa==0.9.1
linkify-it-py==2.0.0
llvmlite==0.40.0
local-attention==1.8.5
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mecab==0.996.3
mecab-python3==1.0.5
mistune==2.0.5
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
Naked==0.1.31
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
num2words==0.5.12
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.2
openai-whisper==20230314
opencv-python==4.7.0.72
orjson==3.8.10
packaging==23.0
pandocfilters==1.5.0
Pandora-ChatGPT==1.1.0
parso==0.8.3
pathos==0.3.0
pathy==0.10.1
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.9.6
platformdirs==2.6.2
playsound==1.3.0
playwright==1.23.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
pooch==1.7.0
portalocker==2.7.0
pox==0.3.2
ppft==1.7.6.6
praat-parselmouth==0.4.3
praw==7.7.0
prawcore==2.3.0
preshed==3.0.8
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
pydub==0.25.1
pyee==8.1.0
PyExecJS==1.5.1
Pygments==2.14.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PyJWT==2.6.0
pynndescent==0.5.8
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pypinyin==0.48.0
pypiwin32==223
pyproject_hooks==1.0.0
pyre-extensions==0.0.23
pyreadline3==3.4.1
pyrsistent==0.19.3
pysbd==0.3.4
PySimpleGUI==4.60.4
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pyttsx3==2.90
pytube==12.1.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
pyworld==0.3.3
PyYAML==6.0
pyzmq==25.0.2
rapidfuzz==2.15.1
readability-lxml==0.8.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
s3transfer==0.6.0
sacrebleu==2.3.1
scikit-learn==1.2.2
scipy==1.9.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentry-sdk==1.17.0
shellescape==3.8.1
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.2.0
sniffio==1.3.0
socksio==1.0.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.26.1
suno-bark @ git+https://github.com/suno-ai/bark.git@97c6019ecdf1a70fdb313835c3bb3f770a0459e4
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.17.1
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.3.1
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
toolz==0.12.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchcrepe==0.0.18
torchvision==0.15.1+cu117
tornado==6.3.2
tqdm==4.65.0
trainer==0.0.20
traitlets==5.9.0
translators==5.3.1
trove-classifiers==2023.5.2
TTS==0.12.0
typer==0.4.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uc-micro-py==1.0.1
umap-learn==0.5.1
Unidecode==1.3.2
unidic-lite==1.0.8
update-checker==0.18.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.21.1
virtualenv==20.21.1
waitress==2.1.2
wasabi==0.10.1
Wave==0.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
win32-setctime==1.1.0
xformers==0.0.18
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
